Small changes for debugging changed Set to List in InputSupplier, so the inputs can be handled in the correct order. Added rudimental implementation of inputs on the ViewImpl class so they can be tested

+++ src/main/java/it/unibo/alienenterprises/model/impl/components/PlayerInputComponentImpl.java
@@ -88,7 +88,7 @@ public class PlayerInputComponentImpl extends ComponentAbs implements PlayerInpu
     @Override
     public void start() {
         this.maxSpeed = getGameObject().getStatValue(Statistic.SPEED);
-        this.acc = this.maxSpeed / ACC_RATE;
+        this.acc = this.maxSpeed / (double) ACC_RATE;
         this.shooter = getGameObject().getComponent(ShooterComponent.class);
     }
